requirements-gpu.txt
numpy
tensorflow-gpu==2.1.0
pandas
tensorflow-datasets
matplotlib


